Renaming kerberos module to kerberos_.
The kerberos module itself imports kerberos, expecting to get the kerberos
package, but it currently gets itself, which means none of the kerberos library
routines are actually available and everything's a non-starter. This follows
pep8's guide for what to do when you have a variable that conflicts with a
builtin. Add a trailing _.